% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmr_omega.R
\name{bmr_omega}
\alias{bmr_omega}
\title{Estimate the effect of sample level variables on background mutation rate}
\usage{
bmr_omega(canno, ganno, l, beta, X, S)
}
\arguments{
\item{canno}{A matix with column Phenotype indicating the number of the silence mutations for every sample}

\item{ganno}{A matrix with column baseline giving position level bmr.}

\item{l}{A vector indicating the lengths of the genes involved}

\item{beta}{A vector giving the coefficients of position level variables}

\item{X}{A data frame giving position level variables}

\item{S}{A data frame giving sample level variables}
}
\value{
A list giving all the information in the model
}
\description{
Estimate the effect of sample level variables on background mutation rate
}
