% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annomut.R
\name{bmrate}
\alias{bmrate}
\title{For each gene and each subject, extract the annotation and mutation data from raw data}
\usage{
bmrate(sgdata, mutation)
}
\arguments{
\item{sgdata}{A list with length 9. Each of its components is a data frame with annotation information. The first eight columns are
chromosome,start,end,ref,alt,genename,functypecode,nttypecode respectively. The other columns
include other position level information.}

\item{mutation}{Data frame with mutation information with columns being
Chromosome, Position, Ref, Alt, SampleID respectively.}
}
\value{
A list named annomut in which length(annomut)=length(genename). Each compoent of annomut
is a list with length equal to the sample size.
}
\description{
For each gene and each subject, extract the annotation and mutation data from raw data
}
