% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmr_signature.R
\name{bmrSignature}
\alias{bmrSignature}
\title{Title}
\usage{
bmrSignature(e, signatures, rho, sc = 1, adjustment = F)
}
\arguments{
\item{e}{phenotype vector}

\item{rho}{Correlation between e and the loadings of first signature.}

\item{sigatures}{A 96 times m matrix M. Each column of it is a signaturee.}

\item{ss}{Scale parameter}
}
\value{
matrix bmr
}
\description{
Title
}
