% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivermaps_readdata.R
\name{convertbeta}
\alias{convertbeta}
\title{Create parameter vector for regular model}
\usage{
convertbeta(j, vbeta)
}
\arguments{
\item{vbeta}{}
}
\value{
A numerical vector
}
\description{
Create parameter vector for regular model
}
