% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivermaps_readdata.R
\name{convertbeta_sig}
\alias{convertbeta_sig}
\title{Create parameter vector for signature model}
\usage{
convertbeta_sig(j, vbeta)
}
\arguments{
\item{vbeta}{}
}
\value{
A numerical vector
}
\description{
Create parameter vector for signature model
}
