% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddmodel.R
\name{dd_EM_ordinary}
\alias{dd_EM_ordinary}
\title{Title}
\usage{
dd_EM_ordinary(
  beta0 = 0,
  alpha = c(0, 0),
  rate.n,
  rate.s0,
  ll.n,
  mutidx,
  type = c("null", "alt"),
  maxit = 100,
  tol = 0.001
)
}
\arguments{
\item{tol}{}
}
\description{
Title
}
