% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivermaps_readdata.R
\name{ddmcode}
\alias{ddmcode}
\title{Add intercept, if have functtypecode, then code and move to the front.
 different from driverMAPS! only allows
 functtypecode =7 ||8 when functypecode is included in selectvars.}
\usage{
ddmcode(matrixlist, selectvars, functypecodelevel = NULL)
}
\description{
Add intercept, if have functtypecode, then code and move to the front.
 different from driverMAPS! only allows
 functtypecode =7 ||8 when functypecode is included in selectvars.
}
