% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddmodel.R
\name{ddmodel_fixb}
\alias{ddmodel_fixb}
\title{diffDriver model with effect size for positional functional annotations fixed}
\usage{
ddmodel_fixb(mut, e, mr, fe)
}
\arguments{
\item{mut}{a matrix of mutation status 0 or 1, rows positions, columns are samples.}

\item{e}{a vector,phenotype of each sample,
should match the columns of \code{mut} and \code{mr}}

\item{mr}{a matrix, mutation rate of each sample at each mutation (log scale) that is not dependent on sample level factor}

\item{fe, }{a vector, increased mutation rate at each position, depending on e (log scale),
should match the rows of \code{mut} and \code{mr}}
}
\description{
This function uses the model as cmodel.frac,
 but generalizes to take more than 1 functional categories.
 This model is applied on data of a single gene
}
