% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivermaps_readdata.R
\name{ddmprocess}
\alias{ddmprocess}
\title{For vars in qnvars, will plug in normal distributed values for missing values using a normal distribution with width of qnvarimpute[2]  and a downshift of the mean by qnvarimpute[1] compared to distribution of all.}
\usage{
ddmprocess(
  matrixlist,
  qnvars = c("expr", "repl", "hic"),
  qnvarimpute = c(-1.8, 0.3),
  cvarimpute = 0,
  fixmusd = NULL
)
}
\arguments{
\item{fixmusd}{}
}
\value{
A list
}
\description{
For vars in qnvars, will plug in normal distributed values for missing values using a normal distribution with width of qnvarimpute[2]  and a downshift of the mean by qnvarimpute[1] compared to distribution of all.
}
