% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivermaps_readdata.R
\name{ddmread}
\alias{ddmread}
\title{Read data}
\usage{
ddmread(
  afileinfo,
  yfileinfo,
  selectvars,
  selectmuttype,
  readinvars = NULL,
  genesubset = NULL
)
}
\arguments{
\item{afileinfo}{Annotation file}

\item{genesubset}{}
}
\value{
A list
}
\description{
Read data
}
