% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivermaps_readdata.R
\name{dtcoldup}
\alias{dtcoldup}
\title{check if any column is duplicate, if yes, stop script
  Note this is a rough way (check sum)}
\usage{
dtcoldup(dt)
}
\arguments{
\item{dt}{}
}
\value{
A warning sign
}
\description{
check if any column is duplicate, if yes, stop script
  Note this is a rough way (check sum)
}
