% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparators.R
\name{genebinom}
\alias{genebinom}
\title{gene level binomial test}
\usage{
genebinom(mut, e)
}
\arguments{
\item{mut}{Mutation data.}

\item{e}{Phenotype or context variable.}
}
\description{
gene level binomial test
}
