% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspot.R
\name{hotspotseq}
\alias{hotspotseq}
\title{Generate the hotspot sequence}
\usage{
hotspotseq(hmm, sgdata)
}
\arguments{
\item{hmm}{transition probability vector}

\item{sgdata}{annotation list}
}
\value{
A n by 2  data frame. The first column start provide the position information, the second column seqt indicates whether this position being a hotspot.
}
\description{
Generate hotspot sequence
}
