% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixlistToBMR.R
\name{matrixlistToBMR}
\alias{matrixlistToBMR}
\title{Title}
\usage{
matrixlistToBMR(adirbase, mutf, BMRlist, k = 5)
}
\arguments{
\item{adirbase}{The path to annotation files}

\item{mutf}{The path to the mutation files}

\item{BMRlist}{The parameters from drivermaps}

\item{target}{The target positions where the bmr's are computed}
}
\description{
Title
}
