% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivermaps_readdata.R
\name{matrixlistToGLM}
\alias{matrixlistToGLM}
\title{Functional mutation}
\usage{
matrixlistToGLM(
  matrixlist,
  chrposmatrixlist,
  BMpars,
  mu_g_s,
  y_g_s,
  fixpars = NULL
)
}
\arguments{
\item{fixpars}{}
}
\value{
A list
}
\description{
Functional mutation
}
