% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivermaps_readdata.R
\name{matrixlistToGLM_sig}
\alias{matrixlistToGLM_sig}
\title{Compute the bmr}
\usage{
matrixlistToGLM_sig(
  matrixlist,
  chrposmatrixlist,
  BMpars,
  mu_g_s,
  y_g_s,
  fixpars = NULL
)
}
\arguments{
\item{fixpars}{}
}
\value{
A list
}
\description{
Compute the bmr
}
