% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivermaps_readdata.R
\name{mlcoluniq}
\alias{mlcoluniq}
\title{check if all column in one dt has at least two values, if not, stop script
   check if all column in different lists have the same set of unique values}
\usage{
mlcoluniq(matrixlist, checkcols = c("functypecode"))
}
\arguments{
\item{checkcols}{}
}
\value{
A list
}
\description{
check if all column in one dt has at least two values, if not, stop script
   check if all column in different lists have the same set of unique values
}
