% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simufun.R
\name{power_compare}
\alias{power_compare}
\title{Simulation function}
\usage{
power_compare(
  family = "binary",
  Niter = 200,
  sgdata,
  bmrpars,
  betaf0 = 0,
  Nsample = 1000,
  beta_gc = c(0, 1.2),
  par
)
}
\arguments{
\item{family}{indicates the type of E variable}

\item{Niter}{is the number of replication in simulation}

\item{sgdata}{is the annotation data}

\item{bmrpars}{is the background mutation rate.}

\item{betaf0}{gene effect mutation}

\item{Nsample}{sample size of simulation}

\item{beta_gc}{the effects of missense and nonsense}

\item{par}{parameters for the simuation}
}
\value{
A list composed of the p-values for 8 models
       and the parameters used in these models.
}
\description{
Simulation function
}
