% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivermaps_readdata.R
\name{readmodeldata}
\alias{readmodeldata}
\title{Read into data}
\usage{
readmodeldata(
  afileinfo,
  yfileinfo,
  selectvars,
  selectmuttype,
  readinvars = NULL,
  qnvars = c("expr", "repl", "hic"),
  functypecodelevel = NULL,
  qnvarimpute = c(-1.8, 0.3),
  cvarimpute = 0,
  genesubset = NULL,
  fixmusd = NULL
)
}
\arguments{
\item{fixmusd}{}
}
\value{
A list
}
\description{
Read into data
}
