% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_cont.R
\name{simulate_2funcv}
\alias{simulate_2funcv}
\title{simulate phenotype and mutations for one gene, using one functional annotation.}
\usage{
simulate_2funcv(sgdata, bmrpars, betaf0 = 2, Nsample, beta_gc, para, hotspot)
}
\arguments{
\item{sgdata}{a list, each list item for one nttype}

\item{bmrpars}{a vector, each item is bmr in log scale for one nttype}

\item{betaf0, }{shift of mutation rate from BMR, log scale, shared in all samples.}

\item{Nsample, }{total number of samples}

\item{beta_gc, }{effect size for functional covariates, log scale. Right now we only have one functional covariate, that is whether the mutation is missense or loss of function mutation. beta_gc[1] indicates the shift of mutation rate for missense ( coded as 7 in functype code column in `sgdata`), beta_gc[2] indicates the shift of mutation rate for loss of function ( coded as 8 in functype code column in `sgdata`). log scale.}

\item{par, }{the parameters for generating the data,  in which par[1] is the mean of phenotype, par[2] the sd of phenotype,
par[3] the intercept of logistic regression and par[4] the slope of the logistic regression. Here the outcome of logistic regression is the
positive sample and the independent variable is the phenotype.}
}
\description{
simulate mutations for one gene. The sample level phenotype are binary (two groups.).
Using only one functional category for position level annotation (missense and loss of function).
}
