% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivermaps_readdata.R
\name{splitddm}
\alias{splitddm}
\title{divide matrixlist into groups defined in cpgenelist,
if not in any group in cpggenelist,
it is put in the last group. need to optimize almost hit 20G}
\usage{
splitddm(matrixlist, cpgenelist)
}
\value{
A list
}
\description{
divide matrixlist into groups defined in cpgenelist,
if not in any group in cpggenelist,
it is put in the last group. need to optimize almost hit 20G
}
