% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivermaps_readdata.R
\name{ddmcode}
\alias{ddmcode}
\title{Add intercept, if have functtypecode, then code and move to the front.}
\usage{
ddmcode(matrixlist, selectvars, functypecodelevel = NULL)
}
\arguments{
\item{functypecodelevel}{}
}
\value{
A list
}
\description{
Add intercept, if have functtypecode, then code and move to the front.
}
