% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddmodel.R
\name{ddmodel}
\alias{ddmodel}
\title{diffDriver model}
\usage{
ddmodel(mut, e, mr, fe, label, ...)
}
\arguments{
\item{mut}{a matrix of mutation status 0 or 1, rows positions, columns are samples.}

\item{e}{a vector,phenotype of each sample,
should match the columns of \code{mut} and \code{mr}}

\item{mr}{a matrix, mutation rate of each sample at each mutation (log scale) that is not dependent on sample level factor}

\item{fe, }{a vector, increased mutation rate at each position, depending on e (log scale),
should match the rows of \code{mut} and \code{mr}}
}
\description{
This model is applied on data of a single gene. It will infer effect size for both sample-level variable and positional level functional annotations. We used an EM algorithm to infer parameters.
}
