% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivermaps_readdata.R
\name{ddmread_j}
\alias{ddmread_j}
\title{Global variable Totalnttype, Outputdir}
\usage{
ddmread_j(fileinfo, j, varlist = NULL, genesubset = NULL)
}
\arguments{
\item{fileinfo}{Anno file information}

\item{j}{the number of nytype}

\item{varlist}{Variables of interest}

\item{genesubset}{Genes of interest}
}
\description{
Global variable Totalnttype, Outputdir
}
