% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffdriver.R
\name{diffDriver}
\alias{diffDriver}
\title{Run diffDriver given input files}
\usage{
diffDriver(genef, mutf, phenof)
}
\arguments{
\item{genef}{file for name of genes to be included in the analysis}

\item{mutf}{mutation list file, use the driverMAPS mutation input format}

\item{phenof}{phenptype file, SampleID <tab> Phenotype <tab> Nsyn. nsyn is number of syn mutations in this sample.}
}
\description{
This is the function to run diffDriver. We first need to set up: run driverMAPS for groups with potential different BMR, assign BMR labels for each sample. then BMR for each sample will be scaled based on driverMAPS results.
}
