% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparators.R
\name{genefisher}
\alias{genefisher}
\title{gene level fisher's exact test}
\usage{
genefisher(mut, e)
}
\arguments{
\item{mut}{Mutation data.}

\item{e}{Phenotype or context variable.}
}
\description{
gene level fisher's exact test
}
