% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparators.R
\name{mlr.v2}
\alias{mlr.v2}
\title{gene level multiple linear regression, correcting for total number of mutations}
\usage{
mlr.v2(mut, e, nmut, covariates = 1)
}
\description{
gene level multiple linear regression, correcting for total number of mutations
}
