% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_mut}
\alias{plot_mut}
\title{plot phenotype, mutation and annotation for a gene across samples}
\usage{
plot_mut(
  gene_name,
  mut,
  pheno,
  totalnttype = 96,
  anno_dir = ".",
  output_prefix = "plot",
  output_dir = "."
)
}
\description{
plot phenotype, mutation and annotation for a gene across samples
}
