% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simufun.R
\name{power_compare}
\alias{power_compare}
\title{Title}
\usage{
power_compare(
  family,
  Niter,
  sgdata,
  Nsample,
  para,
  bmrpars,
  betaf0,
  beta_gc,
  hotseq,
  hmm
)
}
\arguments{
\item{family}{indicates the type of E variable}

\item{Niter}{is the number of replication in simulation}

\item{sgdata}{is the annotation data}

\item{Nsample, }{total number of samples}

\item{para}{para is the parameters for generating the data}

\item{bmrpars}{a vector, each item is bmr in log scale for one nttype}

\item{betaf0, }{shift of mutation rate from BMR, log scale, shared in all samples.}

\item{beta_gc, }{effect size for functional covariates, log scale. Right now we only have one functional covariate, that is whether the mutation is missense or loss of function mutation. beta_gc[1] indicates the shift of mutation rate for missense ( coded as 7 in functype code column in `sgdata`), beta_gc[2] indicates the shift of mutation rate for loss of function ( coded as 8 in functype code column in `sgdata`). log scale.}

\item{...}{other parameters}

\item{hotspot}{hotspot[1] is the probability of being hotspot for a given position. hotspot[2] is the log size
for hotspots.}
}
\value{
A list composed of the p-values for 8 models
and the parameters used in these models.
}
\description{
Title
}
